old-tests/submission/Microsoft/video/video_023.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: Seeking to a value less than 'initialTime' seeks to 'initialTime'.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-seeking" />
  <meta name="assert" content="Seeking to a value less than 'initialTime' seeks to 'initialTime'." />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script type="text/javascript" src="../common/media.js"></script>
  <script type="text/javascript">
  var test = async_test("Seeking to a value less than 'initialTime' seeks to 'initialTime'.")
  
  function getTestVideoElement()
  {
      return document.getElementById("video0");
  }
  
  function do_play(event)
  {
      var vid = getTestVideoElement();
      vid.pause();
      vid.currentTime = -1;
  }
  
  function do_seeked(event)
  {
      var vid = getTestVideoElement();
  
      test.step(function() { assert_equals(vid.initialTime, vid.currentTime)} );
      test.done();
  }
  
  function timeout()
  {
    test.step(function(){ assert_false(true, "'onseeked' event never fired!")});
    test.done();
  }
  
   var timer = setTimeout(timeout, 10000);
  
   </script>
</head>
<body>
  <div id=log></div>
  <script type="text/javascript">
  
  document.write(
      "<video style='display:none' id='video0' src='" + getVideoURI("/media/movie_5") + "'" +
      " onseeked='do_seeked(event);'" +
      " onplay='do_play(event);'" +
      " autoplay>"
      );
  
  </script>
  Your browser does not support media elements.
  </video>
</body>
</html>
